+++ lib/ApiClient/Exception/UnexpectedResponseException.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Ticketpark\ApiClient\Http;
+namespace Ticketpark\ApiClient\Exception;
 
 class UnexpectedResponseException extends \Exception
 {
