+++ test/ApiClient/TicketparkApiClientTest.php
@@ -5,9 +5,9 @@ namespace Ticketpark\ApiClient\Test;
 use PHPUnit\Framework\TestCase;
 use Prophecy\Prophet;
 use Ticketpark\ApiClient\Exception\TokenGenerationException;
+use Ticketpark\ApiClient\Exception\UnexpectedResponseException;
 use Ticketpark\ApiClient\Http\ClientInterface;
 use Ticketpark\ApiClient\Http\Response;
-use Ticketpark\ApiClient\Http\UnexpectedResponseException;
 use Ticketpark\ApiClient\TicketparkApiClient;
 use Ticketpark\ApiClient\Token\AccessToken;
 use Ticketpark\ApiClient\Token\RefreshToken;
